warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@4640 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Storage/VBoxHDD-new.cpp
@@ -495,7 +495,7 @@ VBOXDDU_DECL(int) VDCreate(const char *pszBackend, PFNVDERROR pfnError,
     {
         RTLDRMOD ldrHandle;
         char *pszPluginName;
-        size_t cbPluginName;
+        int cbPluginName;
 
         /* HDD Format Plugins have VBoxHDD as prefix, thatswhy we have to prepend it. 
          * @todo: find out what to do if filenames are case sensitive.
